+++ Library_scheduler/include/CoCoPeLia.hpp
@@ -13,7 +13,7 @@
 #ifndef COCONTROL_H
 #define COCONTROL_H
 typedef struct CoControl{
-	int T = 0;
+	long int T = 0;
 	short dev_num = -1;
 	short dev_ids[LOC_NUM];
 	int Subkernels_per_dev[LOC_NUM];
